+++ examples/bond/test/failing-tests.js
@@ -71,7 +71,8 @@ describe('Bond token failing tests', function () {
 
   this.beforeEach(async function () {
     initialBond = runtime.addAsset(
-      'bond-token-0', { creator: { ...bondTokenCreator.account, name: 'bond-token-creator' } });
+      'bond-token-0',
+      { creator: { ...bondTokenCreator.account, name: 'bond-token-creator' } }).assetID;
 
     const creationFlags = Object.assign({}, flags);
     const creationArgs = [
@@ -85,7 +86,7 @@ describe('Bond token failing tests', function () {
 
     // create application
     applicationId = runtime.addApp(
-      { ...creationFlags, appArgs: creationArgs }, {}, approvalProgram, clearProgram);
+      { ...creationFlags, appArgs: creationArgs }, {}, approvalProgram, clearProgram).appID;
 
     // setup lsig account
     // Initialize issuer lsig with bond-app ID