+++ examples/crowdfunding/test/crowdfundingTest.js
@@ -94,7 +94,7 @@ describe('Crowdfunding Tests', function () {
 
     // create application
     applicationId = runtime.addApp(
-      { ...creationFlags, appArgs: creationArgs }, {}, approvalProgram, clearProgram);
+      { ...creationFlags, appArgs: creationArgs }, {}, approvalProgram, clearProgram).appID;
     const creatorPk = convert.addressToPk(creator.address);
 
     // setup escrow account
@@ -297,7 +297,7 @@ describe('Crowdfunding Tests', function () {
     // create application
     const creationFlags = Object.assign({}, flags);
     const applicationId = runtime.addApp(
-      { ...creationFlags, appArgs: creationArgs }, {}, approvalProgram, clearProgram);
+      { ...creationFlags, appArgs: creationArgs }, {}, approvalProgram, clearProgram).appID;
 
     // setup escrow account
     const lsig = runtime.loadLogic('crowdFundEscrow.py', { APP_ID: applicationId });