+++ examples/permissioned-token/test/common.js
@@ -51,7 +51,7 @@ class Context {
   }
 
   deployASA (name, creator) {
-    this.assetIndex = this.runtime.addAsset(name, { creator: creator });
+    this.assetIndex = this.runtime.addAsset(name, { creator: creator }).assetID;
   }
 
   deployController (sender, approvalProgram, clearStateProgram) {
@@ -67,7 +67,7 @@ class Context {
     const clearProgram = getProgram(clearStateProgram);
     this.controllerappID = this.runtime.addApp(
       sscFlags, {}, controllerProgram, clearProgram
-    );
+    ).appID;
   }
 
   // Deploy Clawback Lsig and Modify Asset
@@ -110,7 +110,7 @@ class Context {
     };
     this.permissionsappID = this.runtime.addApp(
       sscFlags, {}, permissionsProgram, clearProgram
-    );
+    ).appID;
 
     // set permissions SSC app_id in controller ssc
     const appArgs = [