+++ packages/runtime/src/runtime.ts
@@ -9,14 +9,19 @@ import { Ctx } from "./ctx";
 import { RUNTIME_ERRORS } from "./errors/errors-list";
 import { RuntimeError } from "./errors/runtime-errors";
 import { getProgram, Interpreter, loadASAFile } from "./index";
-import { ALGORAND_ACCOUNT_MIN_BALANCE, ALGORAND_MAX_TX_ARRAY_LEN, ZERO_ADDRESS_STR } from "./lib/constants";
+import {
+  ALGORAND_ACCOUNT_MIN_BALANCE, ALGORAND_MAX_TX_ARRAY_LEN, TransactionTypeEnum,
+  ZERO_ADDRESS_STR
+} from "./lib/constants";
 import { convertToString } from "./lib/parsing";
 import { LogicSigAccount } from "./logicsig";
 import { mockSuggestedParams } from "./mock/tx";
 import {
   AccountAddress, AccountStoreI, AppDeploymentFlags, AppOptionalFlags,
   ASADeploymentFlags, ASAInfo, AssetHoldingM, Context,
-  ExecutionMode, RuntimeAccount, SCParams, SSCAttributesM, SSCInfo, StackElem, State, Txn
+  DeployedAppTxReceipt, DeployedAssetTxReceipt,
+  EncTx, ExecutionMode, RuntimeAccount, SCParams, SSCAttributesM, SSCInfo,
+  StackElem, State, TxReceipt
 } from "./types";
 
 export class Runtime {
@@ -44,7 +49,8 @@ export class Runtime {
       assetNameInfo: new Map<string, ASAInfo>(),
       appNameInfo: new Map<string, SSCInfo>(),
       appCounter: ALGORAND_MAX_TX_ARRAY_LEN, // initialize app counter with 8
-      assetCounter: ALGORAND_MAX_TX_ARRAY_LEN // initialize asset counter with 8
+      assetCounter: ALGORAND_MAX_TX_ARRAY_LEN, // initialize asset counter with 8
+      txReceipts: new Map<string, TxReceipt>() // receipt of each transaction, i.e map of {txID: txReceipt}
     };
 
     // intialize accounts (should be done during runtime initialization)
@@ -54,10 +60,18 @@ export class Runtime {
     this.loadedAssetsDefs = loadASAFile(this.store.accountNameAddress);
 
     // context for interpreter
-    this.ctx = new Ctx(cloneDeep(this.store), <Txn>{}, [], [], this);
+    this.ctx = new Ctx(cloneDeep(this.store), <EncTx>{}, [], [], this);
 
     this.round = 2;
     this.timestamp = 1;
+  }
+
+  /**
+   * Returns transaction receipt for a particular transaction
+   * @param txID transaction ID
+   */
+  getTxReceipt (txID: string): TxReceipt | undefined {
+    return this.store.txReceipts.get(txID);
   }
 
   /**
@@ -132,7 +146,7 @@ export class Runtime {
    * Validate first and last rounds of transaction using current round
    * @param gtxns transactions
    */
-  validateTxRound (gtxns: Txn[]): void {
+  validateTxRound (gtxns: EncTx[]): void {
     // https://developer.algorand.org/docs/features/transactions/#current-round
     for (const txn of gtxns) {
       if (Number(txn.fv) >= this.round || txn.lv <= this.round) {
@@ -286,7 +300,7 @@ export class Runtime {
    * @param txnParams : Transaction parameters for current txn or txn Group
    * @returns: [current transaction, transaction group]
    */
-  createTxnContext (txnParams: types.ExecParams | types.ExecParams[]): [Txn, Txn[]] {
+  createTxnContext (txnParams: types.ExecParams | types.ExecParams[]): [EncTx, EncTx[]] {
     // if txnParams is array, then user is requesting for a group txn
     if (Array.isArray(txnParams)) {
       if (txnParams.length > 16) {
@@ -298,7 +312,7 @@ export class Runtime {
         const mockParams = mockSuggestedParams(txnParam.payFlags, this.round);
         const tx = webTx.mkTransaction(txnParam, mockParams);
         // convert to encoded obj for compatibility
-        const encodedTxnObj = tx.get_obj_for_encoding() as Txn;
+        const encodedTxnObj = tx.get_obj_for_encoding() as EncTx;
         encodedTxnObj.txID = tx.txID();
         txns.push(encodedTxnObj);
       }
@@ -308,7 +322,7 @@ export class Runtime {
       const mockParams = mockSuggestedParams(txnParams.payFlags, this.round);
       const tx = webTx.mkTransaction(txnParams, mockParams);
 
-      const encodedTxnObj = tx.get_obj_for_encoding() as Txn;
+      const encodedTxnObj = tx.get_obj_for_encoding() as EncTx;
       encodedTxnObj.txID = tx.txID();
       return [encodedTxnObj, [encodedTxnObj]];
     }
@@ -318,7 +332,7 @@ export class Runtime {
   mkAssetCreateTx (
     name: string, flags: ASADeploymentFlags, asaDef: modelsv2.AssetParams): void {
     // this funtion is called only for validation of parameters passed
-    algosdk.makeAssetCreateTxnWithSuggestedParams(
+    const txn = algosdk.makeAssetCreateTxnWithSuggestedParams(
       flags.creator.addr,
       webTx.encodeNote(flags.note, flags.noteb64),
       asaDef.total,
@@ -336,6 +350,14 @@ export class Runtime {
         : asaDef.metadataHash,
       mockSuggestedParams(flags, this.round)
     );
+
+    if (
+      this.ctx.tx === undefined || this.ctx.tx.type !== TransactionTypeEnum.ASSET_CONFIG
+    ) { // could already be defined (if used as a txGroup in this.executeTx())
+      const encTx = { ...txn.get_obj_for_encoding(), txID: txn.txID() };
+      this.ctx.tx = encTx;
+      this.ctx.gtxs = [encTx];
+    }
   }
 
   /**
@@ -343,12 +365,12 @@ export class Runtime {
    * @param name ASA name
    * @param flags ASA Deployment Flags
    */
-  addAsset (asa: string, flags: ASADeploymentFlags): number {
-    this.ctx.addAsset(asa, flags.creator.addr, flags);
+  addAsset (asa: string, flags: ASADeploymentFlags): DeployedAssetTxReceipt {
+    const txReceipt = this.ctx.addAsset(asa, flags.creator.addr, flags);
     this.store = this.ctx.state;
 
     this.optInToASAMultiple(this.store.assetCounter, this.loadedAssetsDefs[asa].optInAccNames);
-    return this.store.assetCounter;
+    return txReceipt;
   }
 
   /**
@@ -356,12 +378,12 @@ export class Runtime {
    * @param name ASA name
    * @param flags ASA Deployment Flags
    */
-  addASADef (asa: string, asaDef: types.ASADef, flags: ASADeploymentFlags): number {
-    this.ctx.addASADef(asa, asaDef, flags.creator.addr, flags);
+  addASADef (asa: string, asaDef: types.ASADef, flags: ASADeploymentFlags): DeployedAssetTxReceipt {
+    const txReceipt = this.ctx.addASADef(asa, asaDef, flags.creator.addr, flags);
     this.store = this.ctx.state;
 
     this.optInToASAMultiple(this.store.assetCounter, asaDef.optInAccNames);
-    return this.store.assetCounter;
+    return txReceipt;
   }
 
   /**
@@ -387,10 +409,11 @@ export class Runtime {
    * @param address Account address to opt-into asset
    * @param flags Transaction Parameters
    */
-  optIntoASA (assetIndex: number, address: AccountAddress, flags: types.TxParams): void {
-    this.ctx.optIntoASA(assetIndex, address, flags);
+  optIntoASA (assetIndex: number, address: AccountAddress, flags: types.TxParams): TxReceipt {
+    const txReceipt = this.ctx.optIntoASA(assetIndex, address, flags);
 
     this.store = this.ctx.state;
+    return txReceipt;
   }
 
   /**
@@ -449,13 +472,13 @@ export class Runtime {
     flags: AppDeploymentFlags, payFlags: types.TxParams,
     approvalProgram: string, clearProgram: string,
     debugStack?: number
-  ): number {
+  ): DeployedAppTxReceipt {
     this.addCtxAppCreateTxn(flags, payFlags);
     this.ctx.debugStack = debugStack;
-    this.ctx.addApp(flags.sender.addr, flags, approvalProgram, clearProgram, 0);
+    const txReceipt = this.ctx.addApp(flags.sender.addr, flags, approvalProgram, clearProgram, 0);
 
     this.store = this.ctx.state;
-    return this.store.appCounter;
+    return txReceipt;
   }
 
   // creates new OptIn transaction object and update context
@@ -491,12 +514,13 @@ export class Runtime {
    * each opcode execution (upto depth = debugStack)
    */
   optInToApp (accountAddr: string, appID: number,
-    flags: AppOptionalFlags, payFlags: types.TxParams, debugStack?: number): void {
+    flags: AppOptionalFlags, payFlags: types.TxParams, debugStack?: number): TxReceipt {
     this.addCtxOptInTx(accountAddr, appID, payFlags, flags);
     this.ctx.debugStack = debugStack;
-    this.ctx.optInToApp(accountAddr, appID, 0);
+    const txReceipt = this.ctx.optInToApp(accountAddr, appID, 0);
 
     this.store = this.ctx.state;
+    return txReceipt;
   }
 
   // creates new Update transaction object and update context
@@ -544,13 +568,14 @@ export class Runtime {
     payFlags: types.TxParams,
     flags: AppOptionalFlags,
     debugStack?: number
-  ): void {
+  ): TxReceipt {
     this.addCtxAppUpdateTx(senderAddr, appID, payFlags, flags);
     this.ctx.debugStack = debugStack;
-    this.ctx.updateApp(appID, approvalProgram, clearProgram, 0);
+    const txReceipt = this.ctx.updateApp(appID, approvalProgram, clearProgram, 0);
 
     // If successful, Update programs and state
     this.store = this.ctx.state;
+    return txReceipt;
   }
 
   // verify 'amt' microalgos can be withdrawn from account
@@ -602,7 +627,7 @@ export class Runtime {
    * @param to to address
    * @param amount amount of algo in microalgos
    */
-  fundLsig (from: RuntimeAccount, to: AccountAddress, amount: number): void {
+  fundLsig (from: RuntimeAccount, to: AccountAddress, amount: number): TxReceipt {
     const fundParam: types.ExecParams = {
       type: types.TransactionType.TransferAlgo,
       sign: types.SignType.SecretKey,
@@ -611,7 +636,7 @@ export class Runtime {
       amountMicroAlgos: amount,
       payFlags: { totalFee: 1000 }
     };
-    this.executeTx(fundParam);
+    return this.executeTx(fundParam) as TxReceipt;
   }
 
   /**
@@ -620,7 +645,7 @@ export class Runtime {
    * @param debugStack: if passed then TEAL Stack is logged to console after
    * each opcode execution (upto depth = debugStack)
    */
-  validateLsigAndRun (txnParam: types.ExecParams, debugStack?: number): void {
+  validateLsigAndRun (txnParam: types.ExecParams, debugStack?: number): TxReceipt {
     // check if transaction is signed by logic signature,
     // if yes verify signature and run logic
     if (txnParam.sign === types.SignType.LogicSignature && txnParam.lsig) {
@@ -639,6 +664,7 @@ export class Runtime {
         throw new RuntimeError(RUNTIME_ERRORS.GENERAL.INVALID_PROGRAM);
       }
       this.run(program, ExecutionMode.SIGNATURE, 0, debugStack);
+      return this.ctx.state.txReceipts.get(this.ctx.tx.txID) as TxReceipt;
     } else {
       throw new RuntimeError(RUNTIME_ERRORS.GENERAL.LOGIC_SIGNATURE_NOT_FOUND);
     }
@@ -651,7 +677,7 @@ export class Runtime {
    * @param debugStack: if passed then TEAL Stack is logged to console after
    * each opcode execution (upto depth = debugStack)
    */
-  executeTx (txnParams: types.ExecParams | types.ExecParams[], debugStack?: number): void {
+  executeTx (txnParams: types.ExecParams | types.ExecParams[], debugStack?: number): TxReceipt | TxReceipt[] {
     const txnParameters = Array.isArray(txnParams) ? txnParams : [txnParams];
     for (const txn of txnParameters) {
       switch (txn.type) {
@@ -677,10 +703,13 @@ export class Runtime {
 
     // Run TEAL program associated with each transaction and
     // then execute the transaction without interacting with store.
-    this.ctx.processTransactions(txnParameters);
+    const txReceipts = this.ctx.processTransactions(txnParameters);
 
     // update store only if all the transactions are passed
     this.store = this.ctx.state;
+
+    // return transaction receipt(s)
+    return Array.isArray(txnParams) ? txReceipts : txReceipts[0];
   }
 
   /**
@@ -691,14 +720,23 @@ export class Runtime {
    * each opcode execution (upto depth = debugStack)
    * NOTE: Application mode is only supported in TEALv > 1
    */
-  run (program: string, executionMode: ExecutionMode, indexInGroup: number, debugStack?: number): void {
+  run (program: string, executionMode: ExecutionMode,
+    indexInGroup: number, debugStack?: number): TxReceipt {
     const interpreter = new Interpreter();
-    interpreter.execute(program, executionMode, this, debugStack);
+    // set new tx receipt
+    this.ctx.state.txReceipts.set(this.ctx.tx.txID, {
+      txn: this.ctx.tx,
+      txID: this.ctx.tx.txID
+    });
+
     // reset pooled opcode cost for single tx, this is to handle singular functions
     // which don't "initialize" a new ctx (eg. addApp)
     if (this.ctx.gtxs.length === 1) { this.ctx.pooledApplCost = 0; }
+    interpreter.execute(program, executionMode, this, debugStack);
+
     if (executionMode === ExecutionMode.APPLICATION) {
       this.ctx.sharedScratchSpace.set(indexInGroup, interpreter.scratch);
     }
+    return this.ctx.state.txReceipts.get(this.ctx.tx.txID) as TxReceipt;
   }
 }